fixed size table values and size of shift

+++ Code/Core/Inc/structsAndEnums.h
@@ -26,7 +26,7 @@ struct signalInfo{
 	uint32_t signalLocations[waveFormRes] = {0}; //the data array to be written
 	uint16_t frequency = 0; //signal was tested to 80khz, 65k is not enough
 	uint16_t amp = 0; //0-3.3 represented as 0 -4095
-	uint8_t shiftAmount = 0; //shift amount from - 255
+	uint16_t shiftAmount = 0; //shift amount from - 255
 	WaveShape wave = SINE;
 };
